fixing var/www/html

+++ v1/srcs/requirements/wordpress/conf/config.sh
@@ -1,9 +1,21 @@
 #!/bin/bash
-cd /var/www/html
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x wp-cli.phar
-./wp-cli.phar core download --allow-root
-./wp-cli.phar config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=mariadb --allow-root
-./wp-cli.phar core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
 
+# Set up the working directory for WordPress
+cd /var/www/html
+
+# Download WP-CLI if not already present
+if [ ! -f /usr/local/bin/wp ]; then
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    chmod +x wp-cli.phar
+    mv wp-cli.phar /usr/local/bin/wp
+fi
+
+# Download WordPress core files
+if [ ! -f wp-config.php ]; then
+    wp core download --allow-root
+    wp config create --dbname=wordpress --dbuser=wpuser --dbpass=password --dbhost=mariadb --allow-root
+    wp core install --url=localhost --title=inception --admin_user=admin --admin_password=admin --admin_email=admin@admin.com --allow-root
+fi
+
+# Start PHP-FPM
 php-fpm7.4 -F
